11 - restore missing verbs
