--- src/utils.js
@@ -1,5 +1,0 @@
-export function CopyText(text)
-{
-    navigator.clipboard.writeText(text);
-    console.log(text + " copied!");
-}
